+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Use STM based lockless implementation as it results in much better throughput
   in a multi-threaded environment when number of stripes is not equal to the
   number of capabilities (in particular with a single stripe).
+* Stop running resource freeing functions within `uninterruptibleMask`.
 
 # resource-pool-0.4.0.0 (2023-01-16)
 * Require `poolMaxResources` to be not smaller than the number of stripes.